+++ src/cpcli/utils/exceptions.py
